# app/docs/swagger.yaml
definitions:
...
  response.ListData:
    properties:
      rows: {}
      total:
        type: integer
    type: object
  response.ResCode:
    enum:
    - 200
    - 401
    - 403
    - 412
    - 500
    - 600
    type: integer
    x-enum-comments:
      Error: 系统异常
      Forbidden: 没有权限
      Parameter: 参数错误
      Success: 成功
      Unauthorized: token失效
      Waring: 详情看msg
    x-enum-varnames:
    - Success
    - Unauthorized
    - Forbidden
    - Parameter
    - Error
    - Waring
  response.ResponseData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 相应状态码
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
...
  systemModels.SysDictDataVo:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      cssClass:
        type: string
      dictCode:
        example: "0"
        type: string
      dictLabel:
        type: string
      dictSort:
        type: integer
      dictType:
        type: string
      dictValue:
        type: string
      isDefault:
        type: string
      listClass:
        type: string
      remark:
        type: string
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
...
info:
  contact:
    email: zhao_402295440@126.com
    name: danny
  description: baize接口文档
  title: baize
  version: 2.0.x
paths:
...
  /system/config/export:
    post:
      description: 导出字典类型
      parameters:
      - in: query
        name: dictLabel
        type: string
      - in: query
        name: dictType
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      security:
      - BearerAuth: []
      summary: 导出字典类型
      tags:
      - 字典相关
...
  /system/dict/data:
    post:
      description: 添加字典数据
      parameters:
      - description: 字典
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysDictDataVo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 添加字典数据
      tags:
      - 字典相关
    put:
      description: 修改字典数据
      parameters:
      - description: 字典
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysDictDataVo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改字典数据
      tags:
      - 字典相关
  /system/dict/data/{dictCode}:
    get:
      description: 根据dictCode获取字典信息
      parameters:
      - description: dictCode
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysDictDataVo'
              type: object
      security:
      - BearerAuth: []
      summary: 根据dictCode获取字典信息
      tags:
      - 字典相关
  /system/dict/data/{dictCodes}:
    delete:
      description: 根据dictCode获取字典信息
      parameters:
      - collectionFormat: csv
        description: dictCodes
        in: path
        items:
          type: integer
        name: dictCodes
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysDictDataVo'
              type: object
      security:
      - BearerAuth: []
      summary: 根据dictCode获取字典信息
      tags:
      - 字典相关
  /system/dict/data/list:
    get:
      description: 查询字典列表
      parameters:
      - in: query
        name: dictLabel
        type: string
      - in: query
        name: dictType
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysDictDataVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询字典列表
      tags:
      - 字典相关
  /system/dict/data/type/{dictType}:
    get:
      description: 查询字典列表
      parameters:
      - description: dictType
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysDictDataVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询字典列表
      tags:
      - 字典相关
  /system/dict/type:
    delete:
      description: 根据dictCode获取字典类型信息
      parameters:
      - description: 字典
        in: body
        name: object
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 根据dictCode获取字典类型信息
      tags:
      - 字典相关
    post:
      description: 添加字典类型数据
      parameters:
      - description: 字典
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysDictDataVo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 添加字典类型数据
      tags:
      - 字典相关
    put:
      description: 修改字典类型数据
      parameters:
      - description: 字典
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysDictDataVo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改字典类型数据
      tags:
      - 字典相关
  /system/dict/type/{dictCode}:
    get:
      description: 根据dictCode获取字典类型信息
      parameters:
      - description: dictCode
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysDictDataVo'
              type: object
      security:
      - BearerAuth: []
      summary: 根据dictCode获取字典类型信息
      tags:
      - 字典相关
  /system/dict/type/list:
    get:
      description: 查询字典类型列表
      parameters:
      - in: query
        name: dictLabel
        type: string
      - in: query
        name: dictType
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysDictDataVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询字典类型列表
      tags:
      - 字典相关
  /system/dict/type/optionSelect:
    put:
      description: 查询字典列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysDictDataVo'
              type: object
      security:
      - BearerAuth: []
      summary: 查询字典列表
      tags:
      - 字典相关
  /system/dict/type/refreshCache:
    put:
      description: 更新字典缓存
      parameters:
      - description: 字典
        in: body
        name: object
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 更新字典缓存
      tags:
      - 字典相关
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"